# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CervixDetectAI API
  version: 1.0.0
  description: OpenAPI specification for CervixDetectAI backend.
servers:
  - url: /api
    description: Current host (recommended for production)
  - url: https://{domain}/api
    description: Custom domain
    variables:
      domain:
        default: api.example.com
  - url: http://localhost:4000/api
    description: Local dev (4000)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
      required:
        - success
        - message
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
...
    AnalysisTask:
      type: object
      properties:
        id:
          type: integer
        task_id:
          type: string
        study_id:
          type: integer
        user_id:
          type: integer
          nullable: true
        status:
          type: string
          enum: [PENDING, PROCESSING, SUCCESS, FAILED]
        progress:
          type: integer
        ai_model_version:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AnalysisTaskListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                tasks:
                  type: array
                  items:
                    $ref: '#/components/schemas/AnalysisTask'
                pagination:
                  $ref: '#/components/schemas/Pagination'
    CreateBatchAnalysisTaskRequest:
      type: object
      required: [images, patientName, patientId, studyDate, modality]
      properties:
        images:
          type: array
          maxItems: 10
          items:
            type: string
            format: binary
        patientName:
          type: string
        patientId:
          type: string
        studyDate:
          type: string
          format: date
          description: Study date in ISO format
        modality:
          type: string
        description:
          type: string
          nullable: true
        model_version:
          type: string
          nullable: true
        priority:
          type: string
          enum: [normal, urgent, emergency]
    BatchAnalysisTaskItem:
      type: object
      required: [index, originalFilename, status]
      properties:
        index:
          type: integer
        originalFilename:
          type: string
        studyDbId:
          type: integer
          nullable: true
        studyId:
          type: string
          nullable: true
        imageId:
          type: integer
          nullable: true
        taskId:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, FAILED]
        error:
          type: string
          nullable: true
    BatchAnalysisTaskSummary:
      type: object
      required: [total, created, failed]
      properties:
        total:
          type: integer
        created:
          type: integer
        failed:
          type: integer
    BatchAnalysisTaskData:
      type: object
      required: [batchId, summary, items]
      properties:
        batchId:
          type: string
        summary:
          $ref: '#/components/schemas/BatchAnalysisTaskSummary'
        items:
          type: array
          items:
            $ref: '#/components/schemas/BatchAnalysisTaskItem'
    BatchAnalysisTaskResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/BatchAnalysisTaskData'
...
paths:
...
  /analysis-tasks:
    post:
      tags: [AnalysisTasks]
      summary: Create analysis task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [study_id]
              properties:
                study_id:
                  type: integer
                model_name:
                  type: string
                model_version:
                  type: string
                priority:
                  type: string
                  enum: [normal, urgent, emergency]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    get:
      tags: [AnalysisTasks]
      summary: List analysis tasks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: study_id
          schema:
            type: integer
        - in: query
          name: priority
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisTaskListResponse'
  /analysis-tasks/batch:
    post:
      tags: [AnalysisTasks]
      summary: Batch upload images and create analysis tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateBatchAnalysisTaskRequest'
      responses:
        '200':
          description: OK (supports partial success)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchAnalysisTaskResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analysis-tasks/{id}:
    get:
      tags: [AnalysisTasks]
      summary: Get analysis task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    delete:
      tags: [AnalysisTasks]
      summary: Delete analysis task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /analysis-tasks/{id}/status:
    put:
      tags: [AnalysisTasks]
      summary: Update analysis task status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                progress:
                  type: integer
                error_message:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /analysis-tasks/{id}/result:
    post:
      tags: [AnalysisTasks]
      summary: Save analysis result
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [risk_level, diagnosis, confidence]
              properties:
                risk_level:
                  type: string
                diagnosis:
                  type: string
                confidence:
                  type: number
                detailed_report:
                  type: string
                recommendations:
                  type: array
                  items:
                    type: string
                biomarkers:
                  type: object
                suspicious_areas:
                  type: array
                  items:
                    type: object
                raw_output:
                  type: object
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...